Extract object.h and .cc
